Move functions for creating KeyObjects to utility module
