+++ packages/pds/src/index.ts
@@ -31,7 +31,7 @@ import { proxyHandler } from './pipethrough'
 import compression from './util/compression'
 import * as wellKnown from './well-known'
 
-export { createSecretKeyObject } from './auth-verifier'
+export { createSecretKeyObject } from './util/keys'
 export * from './config'
 export { AppContext } from './context'
 export { Database } from './db'
